+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/instruciton_position.go
@@ -1,0 +1,13 @@
+package kurtosis_instruction
+
+type InstructionPosition struct {
+	line int32
+	col  int32
+}
+
+func NewInstructionPosition(line int32, col int32) *InstructionPosition {
+	return &InstructionPosition{
+		line: line,
+		col:  col,
+	}
+}