+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/kurtosis_instruction.go
@@ -5,11 +5,6 @@ import (
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/startosis_engine/startosis_validator"
 )
 
-type InstructionPosition struct {
-	line int32
-	col  int32
-}
-
 type KurtosisInstruction interface {
 	GetPositionInOriginalScript() *InstructionPosition
 
@@ -24,11 +19,4 @@ type KurtosisInstruction interface {
 	// ValidateAndUpdateEnvironment validates if the instruction can be applied to an environment, and mutates that
 	// environment to reflect how Kurtosis would look like after this instruction is successfully executed.
 	ValidateAndUpdateEnvironment(environment *startosis_validator.ValidatorEnvironment) error
-}
-
-func NewInstructionPosition(line int32, col int32) *InstructionPosition {
-	return &InstructionPosition{
-		line: line,
-		col:  col,
-	}
 }