+++ docs/changelog.md
@@ -12,6 +12,7 @@
 * Implemented IP address references for services
 * Added the `defaultTcpLogsCollectorPortNum` with `9713` value in `EngineManager`
 * Added the `LogsCollectorAvailabilityChecker` interface
+* Added an `exec` command to Startosis
 
 ### Changes
 * Add back old enclave continuity test